fixed error with landmark mode with buttons not working
